+++ uti/utiTest/utiTest.vcxproj
@@ -83,6 +83,14 @@
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClInclude Include="..\..\uti.hpp" />
+    <ClInclude Include="..\..\utiAllocator.hpp" />
+    <ClInclude Include="..\..\utiByteIterator.hpp" />
+    <ClInclude Include="..\..\utiCharIterator.hpp" />
+    <ClInclude Include="..\..\utiCommonHeader.hpp" />
+    <ClInclude Include="..\..\utiReferenceCounted.hpp" />
+    <ClInclude Include="..\..\utiReverseIterator.hpp" />
+    <ClInclude Include="..\..\utiUTF16String.hpp" />
+    <ClInclude Include="..\..\utiUTF8String.hpp" />
     <ClInclude Include="stdafx.h" />
     <ClInclude Include="targetver.h" />
   </ItemGroup>
@@ -102,6 +110,15 @@
       <ExcludedFromBuild Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">false</ExcludedFromBuild>
       <DeploymentContent Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">false</DeploymentContent>
     </Text>
+  </ItemGroup>
+  <ItemGroup>
+    <None Include="..\..\utiAllocator.inl" />
+    <None Include="..\..\utiByteIterator.inl" />
+    <None Include="..\..\utiCharIterator.inl" />
+    <None Include="..\..\utiReferenceCounted.inl" />
+    <None Include="..\..\utiReverseIterator.inl" />
+    <None Include="..\..\utiUTF16String.inl" />
+    <None Include="..\..\utiUTF8String.inl" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
